Fix clippy warnings in katla_gfx
- Remove unused imports across Metal backend
- Remove unnecessary unsafe blocks (setLabel is safe in objc2)
- Fix doc comment formatting (empty line, list indentation)
- Fix field_reassign_with_default patterns via clippy --fix

+++ katla_gfx/src/metal/compute_encoder.rs
@@ -74,7 +74,7 @@ impl GpuComputeEncoder<MetalBackend> for MetalComputeEncoder {
         unsafe {
             let ptr = NonNull::new(data.as_ptr() as *mut c_void).unwrap();
             self.inner
-                .setBytes_length_atIndex(ptr, data.len() as usize, index as usize);
+                .setBytes_length_atIndex(ptr, data.len(), index as usize);
         }
     }
